use the preprocessor rather than switch statements to deal with tube type
A switch statement might take memory if the compiler generates a jump table:
http://thegaragelab.com/conserving-memory-on-an-avr/
(although it seems to me that it doesn't, with avr-gcc 4.8, at least).
The tube type will not change at runtime; we only need a multiplier
constant, and the name as a string for concatenation.  The compiler
can do string concatenation too, so there's no reason to use
printf("tube %s") to substitute a string which will never change.
This change saves 30 bytes of RAM.